% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91Data}
\alias{R91Data}
\title{Title for R91Data}
\format{
A data frame with 22 variables:
\describe{
\item{Address}{Address}
\item{MahMorajeh}{The month of visiting the household}
\item{ShoghlSarparast}{The job of the head of the household}
\item{NoeKhn}{Household type (ordinary resident code 1, group code 2)}
\item{Takmil}{Questionnaire for substitute household(Yes cod 1, No code 2)}؟؟؟؟؟؟؟؟؟
\item{TakmilDescA}{Table-2 information on the reason for removing the household(
1.The family does not want to cooperate.
2.The family due to old age, mental disability, illiteracy of all family members or... is not able to answer.
\item{TakmilDescB}{Table-3 information on the reason for replacing the household(
1.The household was absent during the census season.
2.The place is empty.
3.The place is not an ordinary residence for the household.
4.The place is destroyed, under construction or demolition.
5.The residence of the family is an institutional or non-resident.
6.The intend address was not found)}
\item{TakmilDescC}{(Code 7, table-3, other (with the reason for replacing the household)}
\item{Jaygozin}{}   ????????????????
\item{JaygozinDescA}{Table-4 the reason for the removal of the replacement household for reasons of removal)}
\item{JaygozinDescB}{Table-5 the re  ason for the removal of the replacement household for alternative reasons}
\item{JaygozinDescC}{(Code 7, table-5, other (with the reason for the removal of the replacement household for alternative reasons)}
\item{BlkAbdJaygozin}{Block or settlement number of the alternative household (which must be the same as the original sample household)}
\item{RadifJaygozin}{Substitute household row number in the primary sampling unit}
\item{fStat}{Description of var15}
\item{fEnt}{Description of var16}
\item{fvfy}{Description of var17}
\item{user1}{Description of var18}
\item{user2}{Description of var19}
\item{user3}{Description of var20}
\item{user4}{Description of var21}
\item{Tekrari}{Description of var22}
}
}
\usage{
R91Data
}
\description{
Title for R91Data
}
\keyword{datasets}
