% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91KarBargh2}
\alias{R91KarBargh2}
\title{Title for R91KarBargh2
Income (employees in the agricultural sector) of Rural Households in 1391
كاربرگ مربوط به درامدهاي متفرقه اعضاء خانوار سال 1391}
\format{
A data frame with 95 variables:
\describe{Income (employees in the agricultural sector)
\item{Address}{"آدرس" Address}
\item{DYCOL01}{"شماره ردیف فرد" Person row number}
\item{DYCOL02}{"کد شغل" Job code}
\item{DYCOL03}{"نوع و مشخصات فعاليت ها(زراعت)"
"Type and feature of activities (agriculture)"
classified in DyCole04 - DyCole14}
\item{DYCOL04}{"سطح زير كشت" Related to agricultural activities}
\item{DYCOL05}{"باغداری" Related to agricultural activities}
\item{DYCOL06}{"مساحت باغ" Related to agricultural activities}
\item{DYCOL07}{"دامداری" Related to agricultural activities}
\item{DYCOL08}{"تعداد دام کوچک" Related to agricultural activities}
\item{DYCOL09}{"تعداد دام بزرگ" Related to agricultural activities}
\item{DYCOL10}{"مرغداری" Related to agricultural activities}
\item{DYCOL11}{"تعداد مرغ، جوجه و خروس" Related to agricultural activities}
\item{DYCOL12}{"پرورش ماهی" Related to agricultural activities}
\item{DYCOL13}{"مساحت استخر" Related to agricultural activities}
\item{DYCOL14}{"سایر فعالیت های کشاورزی" Related to agricultural activities}
\item{DYCOL15}{"مزد و حقوق و مزايا (پولي و جنسي )دستمزد ميراب"
"Wages, salaries and benefits (monetary and goods)"
classified in DyCole16 - DyCole25}
\item{DYCOL16}{("دستمزد دشتبان") Wages, salaries and benefits (monetary and goods)}
\item{DYCOL17}{"دستمزد دروگر" Wages, salaries and benefits (monetary and goods)}
\item{DYCOL18}{"دستمزد وجین کار" Wages, salaries and benefits (monetary and goods)}
\item{DYCOL19}{"دستمزد چوپان" Wages, salaries and benefits (monetary and goods)}
\item{DYCOL20}{"دستمزد چوبکش" Wages, salaries and benefits (monetary and goods)}
\item{DYCOL21}{"دستمزد میوه چین" Wages, salaries and benefits (monetary and goods)}
\item{DYCOL22}{"دستمزد باربر" Wages, salaries and benefits (monetary and goods)}
\item{DYCOL23}{"دستمزد باغبان" Wages, salaries and benefits (monetary and goods)}
\item{DYCOL24}{"سایر دستمزدهای پرداختی" Other wages, salaries and benefits (monetary and goods)}
\item{DYCOL25}{"جمع هزینه ها" Total wages, salaries and benefits (monetary and goods)}
\item{DYCOL26}{"بذر، كود، سم، علوفه، هزينه سوخت و تعميرات لوازم كار و نظاير آن- تهيه دام براي پروار بندي"
"Seeds, fertilizers, poisons, fodder, fuel costs and equipment repairs Work and the like - preparing livestock for fattening"
Classified in DyCole27 - DyCole45}
\item{DYCOL27}{"تهیه انواع بذر و نهال" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL28}{"تهیه کود شیمیایی و ویتامین های مختلف" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL29}{"تهیه کود حیوانی از محل اشتغال دامداری" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL30}{"کود حیوانی خریداری شده" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL31}{"کود حیوانی خریداری شده از محل اشتغال" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL32}{"سوخت تراکتور کمباین" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL33}{"سوخت تیلر وسایل و سایر وسایل نقلیه" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL34}{"تعمیرات موتور آب" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL35}{"تعمیرات تراکتور، کمباین" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL36}{"تعمیرات خرمنکوب، تیلر" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL37}{"تعمیرات سایر وسایل نقلیه" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL38}{"سوخت موتور آب" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL39}{"تهیه سم خریداری شده و سمپاش" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL40}{"تیز کردن ابزار کار" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL41}{"تهیه بنزین و گازويیل وسایل نقلیه" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL42}{"تهیه علوفه و کاه" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL43}{"شکر، موم و قاب برای زنبور عسل" Agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL44}{"سایر هزینه های مشابه" Other agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL45}{"جمع هزینه ها" Total agri-Services: Inputs, Maintenance, Fuel}
\item{DYCOL46}{"تهيه ابزار كار بي دوام و هزينه استهلاك ابزار و ماشين آلات بادوام – سطل"
"Provision of non-durable work tools and cost of consumption of durable tools and machinery - bucket"
Classified in DyCole47 - DyCole54}
\item{DYCOL47}{"دبه پلاستیکی" Costs and consumables for tool procurement & maintenance}
\item{DYCOL48}{"استهلاک موتور آب" Costs and consumables for tool procurement & maintenance}
\item{DYCOL49}{"استهلاک تیلر و خرمنکوب" Costs and consumables for tool procurement & maintenance}
\item{DYCOL50}{"استهلاک وسایل نقلیه شغل زراعت" Costs and consumables for tool procurement & maintenance}
\item{DYCOL51}{"استهلاک سایر وسایل عمده کار" Costs and consumables for tool procurement & maintenance}
\item{DYCOL52}{"تعمیرات غیراساسی" Costs and consumables for tool procurement & maintenance}
\item{DYCOL53}{"سایر" Other costs and consumables for tool procurement & maintenance}
\item{DYCOL54}{"جمع هزینه ها" Total costs and consumables for tool procurement & maintenance}
\item{DYCOL55}{"كارمزد وام شغلي ، اجاره ابزار و زمين ، حمل و نقل و بسته بندي و نظاير آن- كارمزد وام شغلي"
"Job loan fee, rent of tools and land, transportation, Packaging and the like - Job loan fee"
Classified in DyCole56 - DyCole67}
\item{DYCOL56}{"اجاره تراکتور(باراننده و بدون راننده)" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL57}{"اجاره کمباین" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL58}{"اجاره خرمنکوب و تیلر" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL59}{"اجاره گاوکار و خیش" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL60}{"اجاره زمین و باغ(نقدی و کالایی)" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL61}{"حمل و نقل محصول" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL62}{"تهیه گونی، جعبه، کیسه نایلونی، طناب، کارتن مقوایی و غیره" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL63}{"لایروبی چاه" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL64}{"آب" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL65}{"برق" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL66}{"سایر هزینه های مشابه" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL67}{"جمع هزینه ها" Operating Costs: Job Fees, Rent, Transportation, Packaging}
\item{DYCOL68}{"محصولات کشاورزی" "Agricultural products"
Classified in DyCole69 - DyCole90}
\item{DYCOL69}{"کاه و علوفه" Agricultural products}
\item{DYCOL70}{"محصولات جالیزی" Agricultural products}
\item{DYCOL71}{"محصولات صیفی" Agricultural products}
\item{DYCOL72}{"سبزی های برگی" Agricultural products}
\item{DYCOL73}{"سبزی های بوته ای" Agricultural products}
\item{DYCOL74}{"سبزی های ریشه ای" Agricultural products}
\item{DYCOL75}{"میوه های درختی(بجز مرکبات)" Agricultural products}
\item{DYCOL76}{"محصولات درختی(مرکبات)" Agricultural products}
\item{DYCOL77}{"اصله و نهال" Agricultural products}
\item{DYCOL78}{"خاک از بذر گل ها" Agricultural products}
\item{DYCOL79}{"گل" Agricultural products}
\item{DYCOL80}{"آب" Agricultural products}
\item{DYCOL81}{"لبنیات" Agricultural products}
\item{DYCOL82}{"پشم، مو، کرک و پر مرغ" Agricultural products}
\item{DYCOL83}{"کودهای حیوانی" Agricultural products}
\item{DYCOL84}{"انواع دام" Agricultural products}
\item{DYCOL85}{"جوجه های تولیدی" Agricultural products}
\item{DYCOL86}{"زنبورهای تولیدی" Agricultural products}
\item{DYCOL87}{"عسل و موم" Agricultural products}
\item{DYCOL88}{"تخم مرغ" Agricultural products}
\item{DYCOL89}{"سایر اغلام مشابه" Agricultural products}
\item{DYCOL90}{"جمع فروش" Agricultural products}
}
}
\usage{
R91KarBargh2
}
\description{
Title for R91KarBargh2
Income (employees in the agricultural sector) of Rural Households in 1391
كاربرگ مربوط به درامدهاي متفرقه اعضاء خانوار سال 1391
}
\keyword{datasets}
