% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91P2}
\alias{R91P2}
\title{Table R91P2
Household Facilities in the city in 1391
مشخصات محل سكونت و تسهيلات و لوازم عمده زندگي شهری در سال 1391}
\format{
A data frame with 62 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{نحوه ی تصرف محل سکونت}
\item Description: tenure
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{ساير انواع نحوه ي تصرف}
\item Description: Other tenure
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{تعداد اتاق در اختيار خانوار}
\item Description: Room
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{سطح زيربنا}
\item Description: Space
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{نوع اسكلت بناي محل سكونت}
\item Description: Construction
}
\item \code{DYCOL06}:
\itemize{
\item Farsi name: \emph{مصالح عمده}
\item Description: Material
}
\item \code{DYCOL07}:
\itemize{
\item Farsi name: \emph{اتومبيل سواري شخصي}
\item Description: Vehicle
}
\item \code{DYCOL08}:
\itemize{
\item Farsi name: \emph{موتورسيكلت}
\item Description: Motorcycle
}
\item \code{DYCOL09}:
\itemize{
\item Farsi name: \emph{دوچرخه}
\item Description: Bicycle
}
\item \code{DYCOL10}:
\itemize{
\item Farsi name: \emph{راديو}
\item Description: Radio
}
\item \code{DYCOL11}:
\itemize{
\item Farsi name: \emph{راديو ضبط، ضبط و پخش صوت}
\item Description: Radio tape
}
\item \code{DYCOL12}:
\itemize{
\item Farsi name: \emph{تلوزيون (سياه وسفيد)}
\item Description: TV (black and white)
}
\item \code{DYCOL13}:
\itemize{
\item Farsi name: \emph{تلويزيون رنگي}
\item Description: TV
}
\item \code{DYCOL14}:
\itemize{
\item Farsi name: \emph{VCD و DVD انواع ویذیو و}
\item Description: VHS_VCD_DVD
}
\item \code{DYCOL15}:
\itemize{
\item Farsi name: \emph{رایانه}
\item Description: Computer
}
\item \code{DYCOL16}:
\itemize{
\item Farsi name: \emph{تلفن همراه (غير شغلي)}
\item Description: Cellphone
}
\item \code{DYCOL17}:
\itemize{
\item Farsi name: \emph{فريزر}
\item Description: Freezer
}
\item \code{DYCOL18}:
\itemize{
\item Farsi name: \emph{يخچال}
\item Description: Refrigerator
}
\item \code{DYCOL19}:
\itemize{
\item Farsi name: \emph{يخچال فريزر}
\item Description: Fridge
}
\item \code{DYCOL20}:
\itemize{
\item Farsi name: \emph{اجاق گاز}
\item Description: Stove
}
\item \code{DYCOL21}:
\itemize{
\item Farsi name: \emph{جاروبرقي}
\item Description: Vacuum
}
\item \code{DYCOL22}:
\itemize{
\item Farsi name: \emph{ماشین لباسشویی}
\item Description: Washing machine
}
\item \code{DYCOL23}:
\itemize{
\item Farsi name: \emph{چرخ خياطي}
\item Description: Sewingmachine
}
\item \code{DYCOL24}:
\itemize{
\item Farsi name: \emph{پنکه}
\item Description: Fan
}
\item \code{DYCOL25}:
\itemize{
\item Farsi name: \emph{کولر آبي متحرک}
\item Description: Evap cooling portable
}
\item \code{DYCOL26}:
\itemize{
\item Farsi name: \emph{كولر گازي متحرك}
\item Description: Split portable
}
\item \code{DYCOL27}:
\itemize{
\item Farsi name: \emph{ماشين ظرفشويي}
\item Description: Dishwasher
}
\item \code{DYCOL28}:
\itemize{
\item Farsi name: \emph{مايكرو ويو و فرهاي هالوژن دار}
\item Description: Microwave and halogen ovens
}
\item \code{DYCOL29}:
\itemize{
\item Farsi name: \emph{هيچكدام}
\item Description: None
}
\item \code{DYCOL30}:
\itemize{
\item Farsi name: \emph{آب لوله كشي}
\item Description: Pipewater
}
\item \code{DYCOL31}:
\itemize{
\item Farsi name: \emph{برق}
\item Description: Electricity
}
\item \code{DYCOL32}:
\itemize{
\item Farsi name: \emph{گازلوله كشي}
\item Description: Pipegas
}
\item \code{DYCOL33}:
\itemize{
\item Farsi name: \emph{تلفن}
\item Description: Telephone
}
\item \code{DYCOL34}:
\itemize{
\item Farsi name: \emph{اينترنت}
\item Description: Internet
}
\item \code{DYCOL35}:
\itemize{
\item Farsi name: \emph{حمام}
\item Description: Bathroom
}
\item \code{DYCOL36}:
\itemize{
\item Farsi name: \emph{آشپزخانه}
\item Description: Kitchen
}
\item \code{DYCOL37}:
\itemize{
\item Farsi name: \emph{كولرآبي ثابت}
\item Description: Evap cooling
}
\item \code{DYCOL38}:
\itemize{
\item Farsi name: \emph{برودت مرکزی}
\item Description: Central cooling
}
\item \code{DYCOL39}:
\itemize{
\item Farsi name: \emph{حرارت مركزی}
\item Description: Centralheating
}
\item \code{DYCOL40}:
\itemize{
\item Farsi name: \emph{پكيج}
\item Description: Package
}
\item \code{DYCOL41}:
\itemize{
\item Farsi name: \emph{كولر گازي ثابت}
\item Description: Split
}
\item \code{DYCOL42}:
\itemize{
\item Farsi name: \emph{شبكه فاضلاب شهري}
\item Description: Waste water
}
\item \code{DYCOL43}:
\itemize{
\item Farsi name: \emph{نوع سوخت عمده مصرفی خانوار (پخت و پز)}
\item Description: Cooking fuel
}
\item \code{DYCOL44}:
\itemize{
\item Farsi name: \emph{نوع سوخت عمده مصرفی خانوار (گرما)}
\item Description: Heating fuel
}
\item \code{DYCOL45}:
\itemize{
\item Farsi name: \emph{نوع سوخت عمده مصرفی خانوار (تهيه آب گرم )}
\item Description: Waterheatingfuel
}
\item \code{DYCOL46}:
\itemize{
\item Farsi name: \emph{جشن ها- ماه گذشته}
\item Description: Celebrations - last month
}
\item \code{DYCOL47}:
\itemize{
\item Farsi name: \emph{جشن ها- سال گذشته}
\item Description: Celebrations - last year
}
\item \code{DYCOL48}:
\itemize{
\item Farsi name: \emph{عزاداري ها- ماه گذشته}
\item Description: Mourning - last month
}
\item \code{DYCOL49}:
\itemize{
\item Farsi name: \emph{عزاداريها- سال گذشته}
\item Description: Mourning - last year
}
\item \code{DYCOL50}:
\itemize{
\item Farsi name: \emph{تعميرات منزل مسكوني- ماه گذشته}
\item Description: Residential home repairs - last month
}
\item \code{DYCOL51}:
\itemize{
\item Farsi name: \emph{تعميرات منزل مسكوني- سال گذشته}
\item Description: Residential home repairs - last year
}
\item \code{DYCOL52}:
\itemize{
\item Farsi name: \emph{سفرهاي زيارتي- ماه گذشته}
\item Description: Pilgrimage trips - last month
}
\item \code{DYCOL53}:
\itemize{
\item Farsi name: \emph{سفرهاي زيارتي- سال گذشته}
\item Description: Pilgrimage trips - last year
}
\item \code{DYCOL54}:
\itemize{
\item Farsi name: \emph{ساير سفرهاي خارجي- ماه گذشته}
\item Description: Other foreign trips - last month
}
\item \code{DYCOL55}:
\itemize{
\item Farsi name: \emph{ساير سفرهاي خارجي- سال گذشته}
\item Description: Other foreign trips - last year
}
\item \code{DYCOL56}:
\itemize{
\item Farsi name: \emph{عمل هاي جراحي- ماه گذشته}
\item Description: Surgeries - last month
}
\item \code{DYCOL57}:
\itemize{
\item Farsi name: \emph{عمل هاي جراحي- سال گذشته}
\item Description: Surgeries - last year
}
\item \code{DYCOL58}:
\itemize{
\item Farsi name: \emph{ساير مناسبت ها- ماه گذشته}
\item Description: Other occasions - last month
}
\item \code{DYCOL59}:
\itemize{
\item Farsi name: \emph{ساير مناسبت ها- سال گذشته}
\item Description: Other occasions - last year
}
\item \code{DYCOL60}:
\itemize{
\item Farsi name: \emph{نام ساير مناسبت ها}
\item Description: Names of other occasions
}
\item \code{DYCOL61}:
\itemize{
\item Farsi name: \emph{هيچكدام}
\item Description: None
}
}

A data frame with 62 variables:
\describe{"جدول قسمت دوم، مشخصات محل سكونت و تسهيلات و لوازم عمده زندگي" The table of the second part, the details of the place of residence and the facilities and major necessities of life
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"نحوه تصرف محل سكونت" The details of the residency and the facilities}
\item{DYCOL02}{"ساير انواع نحوه تصرف" The details of the residency and the facilities}
\item{DYCOL03}{"تعداد اتاق در اختیار خانواده" The details of the residency and the facilities}
\item{DYCOL04}{"سطح زیربنا" The details of the residency and the facilities}
\item{DYCOL05}{"نوع اسکلت بنای محل سکونت" The details of the residency and the facilities}
\item{DYCOL06}{"مصالح عمده" The details of the residency and the facilities}
\item{DYCOL07}{"اتومبیل سواری شخصی" The major necessities of life}
\item{DYCOL08}{"موتورسیکلت" The major necessities of life}
\item{DYCOL09}{"دوچرخه" The major necessities of life}
\item{DYCOL10}{"رادیو" The major necessities of life}
\item{DYCOL11}{"رادیو ضبط، ضبط و پخش صوت" The major necessities of life}
\item{DYCOL12}{"تلویزیون(سیاه و سفید)" The major necessities of life}
\item{DYCOL13}{"تلویزیون رنگی" The major necessities of life}
\item{DYCOL14}{"انواع ویدئو VCD و DVD" The major necessities of life}
\item{DYCOL15}{"رایانه" The major necessities of life}
\item{DYCOL16}{"تلفن همراه (غیرشغلی)" The major necessities of life}
\item{DYCOL17}{"فریزر" The major necessities of life}
\item{DYCOL18}{"یخچال" The major necessities of life}
\item{DYCOL19}{"یخچال فریزر" The major necessities of life}
\item{DYCOL20}{"اجاق گاز" The major necessities of life}
\item{DYCOL21}{"جاروبرقی" The major necessities of life}
\item{DYCOL22}{"ماشین لباسشویی" The major necessities of life}
\item{DYCOL23}{"چرخ خیاطی" The major necessities of life}
\item{DYCOL24}{"پنکه" The major necessities of life}
\item{DYCOL25}{"کولر آبی متحرک" The major necessities of life}
\item{DYCOL26}{"کولر گازی متحرک" The major necessities of life}
\item{DYCOL27}{"ماشین ظرفشویی" The major necessities of life}
\item{DYCOL28}
\item{DYCOL29}{"هیچکدام" The major necessities of life}
\item{DYCOL30}{"آب لوله کشی" The major necessities of life}
\item{DYCOL31}{"برق" The major necessities of life}
\item{DYCOL32}{"گاز لوله کشی" The major necessities of life}
\item{DYCOL33}{"تلفن" The major necessities of life}
\item{DYCOL34}{"اینترنت" The major necessities of life}
\item{DYCOL35}{"حمام" The major necessities of life}
\item{DYCOL36}{"آشپزخانه" The major necessities of life}
\item{DYCOL37}{"کولر آبی ثابت" The major necessities of life}
\item{DYCOL38}{"برودت مرکزی" The major necessities of life}
\item{DYCOL39}{"حرارت مرکزی" The major necessities of life}
\item{DYCOL40}{"پکیج" The major necessities of life}
\item{DYCOL41}{"کولر گاز ثابت" The major necessities of life}
\item{DYCOL42}{"شبکه فاضلاب شهری" The major necessities of life}
\item{DYCOL43}{"نوع سوخت عمده مصرفي خانوار(پخت و پز)" The major necessities of life}
\item{DYCOL44}{"نوع سوخت عمده مصرفي خانوار(گرما)" The major necessities of life}
\item{DYCOL45}{"نوع سوخت عمده مصرفي خانوار(تهيه آب گرم )"The major necessities of life}
\item{DYCOL46}{"جشن ها-ماه گذشته" The major necessities of life}
\item{DYCOL47}{"جشن ها-سال گذشته" The major necessities of life}
\item{DYCOL48}{"عزاداری ها-ماه گذشته" The major necessities of life}
\item{DYCOL49}{"عزاداری ها-سال گذشته" The major necessities of life}
\item{DYCOL50}{"تعمیرات منزل مسکونی-ماه گذشته" The major necessities of life}
\item{DYCOL51}{"تعمیرات منزل مسکونی-سال گذشته" The major necessities of life}
\item{DYCOL52}{"سفرهای زیارتی-ماه گذشته" The major necessities of life}
\item{DYCOL53}{"سفرهای زیارتی-سال گذشته" The major necessities of life}
\item{DYCOL54}{"سایر سفرهای خارجی-ماه گذشته" The major necessities of life}
\item{DYCOL55}{"سایر سفرهای خارجی-سال گذشته" The major necessities of life}
\item{DYCOL56}{"عمل های جراحی-ماه گذشته" The major necessities of life}
\item{DYCOL57}{"عمل های جراحی-سال گذشته" The major necessities of life}
\item{DYCOL58}{"سایر مناسبت ها-ماه گذشته" The major necessities of life}
\item{DYCOL59}{"سایر مناسبت ها-سال گذشته" The major necessities of life}
\item{DYCOL60}{"هیچکدام" The major necessities of life}
\item{DYCOL61}{}
}
}
\usage{
R91P2

R91P2
}
\description{
{R for Rural} \cr
{91 for the year of 1391}
}
\keyword{datasets}
