% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91P3S01}
\alias{R91P3S01}
\title{Title for R91P3S01}
\format{
A data frame with 7 variables:
\describe{"هزينه هائ خوراكي ودخاني" Food and tobacco expenses
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"کد کالا" Bar code}
\item{DYCOL02}{"طریق تهیه" Preparation method}
\item{DYCOL03}{"مقدار گرم" Amount of grams}
\item{DYCOL04}{"مقدار کیلو" Amount of kilogram}
\item{DYCOL05}{"قیمت واحد(ریال)" Unit price (Rials)}
\item{DYCOL06}{"ارزش(ریال)" Value (Rials)}
}
}
\usage{
R91P3S01
}
\description{
Title for R91P3S01
}
\keyword{datasets}
