% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91P3S09}
\alias{R91P3S09}
\title{Title for R91P3S09}
\format{
A data frame with 4 variables:
\describe{"هزینه های مربوط به خدمات فرهنگی و تفریحات خانوار" Expenses related to cultural services and household entertainment
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"کد کالا" Bar code}
\item{DYCOL02}{"طریق تهیه" Preparation method}
\item{DYCOL03}{"ارزش(ریال)" Value (Rials)}
}
}
\usage{
R91P3S09
}
\description{
Title for R91P3S09
}
\keyword{datasets}
