% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91P4S03}
\alias{R91P4S03}
\title{Title for R91P4S03}
\format{
A data frame with 8 variables:
\describe{"درآمدهای متفرقه خانوار" Miscellaneous household income
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"شماره ردیف اعضای خانوار" Row number of family members}
\item{DYCOL03}{"حقوق بازنشستگی" Retirement wage}
\item{DYCOL04}{"درآمد حاصل از اجاره محل کسب" Revenue from renting the business place}
\item{DYCOL05}{"درآمد حاصل از حساب پس انداز" Income from the savings account}
\item{DYCOL06}{"کمک هزینه تحصیلی" Scholarships}
\item{DYCOL07}{"درآمد حاصل از محل فروش مصنوعات" Income from the sale of artifacts}
\item{DYCOL08}{"دریافتی های انتقالی از خانوارهای دیگر" Transfer receivals from other families}
}
}
\usage{
R91P4S03
}
\description{
Title for R91P4S03
}
\keyword{datasets}
