% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91KarBargh1}
\alias{U91KarBargh1}
\title{Title for U91KarBargh1}
\format{
A data frame with 4 variables:
\describe{
\item{Address}{"آدرس" Address}
\item{DYCOL01}{"شماره ردیف فرد" Person row number}
\item{DYCOL02}{"شماره ردیف فعالیت" Activity row number}
\item{DYCOL03}{"درآمد" Income}
}
}
\usage{
U91KarBargh1
}
\description{
Title for U91KarBargh1
}
\keyword{datasets}
