% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91P1}
\alias{U91P1}
\title{Title for U91P1}
\format{
A data frame with 10 variables:
\describe{"جدول قسمت يكم، خصوصيات اجتماعي اعضاي خانوار" The table of the first part, The social characteristics of the family members
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"شماره ردیف اعضای خانوار" Row number of family members}
\item{DYCOL03}{"بستگی با سرپرست" Relationship with the guardian}
\item{DYCOL04}{"جنس" Gender}
\item{DYCOL05}{"سن" Age}
\item{DYCOL06}{"وضع سواد" Literacy status}
\item{DYCOL07}{"آیا در حال حاضر تحصیل می کند؟" Is she/he currently studying?}
\item{DYCOL08}{"پایه یا مدرک" Degree}
\item{DYCOL09}{"وضع فعالیت" Career status}
\item{DYCOL10}{"وضع زناشویی" Marital status}
}
}
\usage{
U91P1
}
\description{
Title for U91P1
}
\keyword{datasets}
