% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91P3S04}
\alias{U91P3S04}
\title{Title for U91P3S04}
\format{
A data frame with 5 variables:
\describe{"هزينه های پوشاک" Cloth expenses
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"کد کالا" Bar code}
\item{DYCOL02}{"تعداد افراد بیمه شده" Number of insured persons}
\item{DYCOL03}{"طریق تهیه" Preparation method}
\item{DYCOL04}{"ارزش(ریال)" Value (Rials)}
}
}
\usage{
U91P3S04
}
\description{
Title for U91P3S04
}
\keyword{datasets}
