% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91P3S06}
\alias{U91P3S06}
\title{Table U91P3S06 \cr
Household health & hygiene expenditure of urban household in 1391 \cr
هزينه هاي  بهداشتي ودرماني خانوار شهری سال 1391}
\format{
A data frame with 4 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس خانوار}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
U91P3S06
}
\description{
{U for Urban} \cr
{91 for 1391} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
