% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91P3S07}
\alias{U91P3S07}
\title{Title for U91P3S07}
\format{
A data frame with 4 variables:
\describe{"هزینه های جاری حمل و نقل" Current transportation costs
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"کد کالا" Bar code}
\item{DYCOL02}{"طریق تهیه" Preparation method}
\item{DYCOL03}{"ارزش(ریال)" Value (Rials)}
}
}
\usage{
U91P3S07
}
\description{
Title for U91P3S07
}
\keyword{datasets}
