% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91P3S12}
\alias{U91P3S12}
\title{Title for U91P3S12}
\format{
A data frame with 4 variables:
\describe{"هزینه های کالاها و خدمات متفرقه خانوار" Expenses of miscellaneous household goods and services
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"کد کالا" Bar code}
\item{DYCOL02}{"طریق تهیه" Preparation method}
\item{DYCOL03}{"ارزش(ریال)" Value (Rials)}
}
}
\usage{
U91P3S12
}
\description{
Title for U91P3S12
}
\keyword{datasets}
