% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91P3S13}
\alias{U91P3S13}
\title{Title for U91P3S13}
\format{
A data frame with 7 variables:
\describe{"تهیه و فروش کالاهای بادوام منزل" Preparation and sale of durable household goods
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"کد کالا" Bar code}
\item{DYCOL02}{"تعداد افراد بیمه شده" Number of insured persons}
\item{DYCOL03}{"طریق تهیه" Preparation method}
\item{DYCOL04}{"خرید یا هزینه" Purchase or expense}
\item{DYCOL05}{"فروش دست دوم" Second hand sale}
}
}
\usage{
U91P3S13
}
\description{
Title for U91P3S13
}
\keyword{datasets}
