% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91P3S14}
\alias{U91P3S14}
\title{Title for U91P3S14}
\format{
A data frame with 7 variables:
\describe{"سرمایه گذاری خانوار در 12 ماه گذشته" Household investment in the last 12 months
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"کد کالا" Bar code}
\item{DYCOL02}{"طریق تهیه" Preparation method}
\item{DYCOL03}{"خرید یا هزینه" Purchase or expense}
\item{DYCOL04}{"فروش دست دوم" Second hand sale}
}
}
\usage{
U91P3S14
}
\description{
Title for U91P3S14
}
\keyword{datasets}
