% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91}
\alias{R91}
\title{Title for R91}
\format{
A data frame with 2 variables:
\describe{
\item{Address}{Address}
\item{MahMorajeh}{The month of visiting the household}
}
}
\usage{
R91
}
\description{
Title for R91
}
\keyword{datasets}
