% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91}
\alias{R91}
\title{Table R91}
\format{
A data frame with 2 columns
\describe{
\item {Address:} {"آدرس"} {" Household address"}
\item {MahMorajeh:} {"ماه مراجعه به خانوار"} {"The month of visiting the household"}
}
}
\usage{
R91
}
\description{
R for Rural
91 for year of 1391
}
\keyword{datasets}
