% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91}
\alias{R91}
\title{Table R91}
\format{
A data frame with 2 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{MahMorajeh}:
\itemize{
\item Farsi name: \emph{ماه مراجعه به خانوار}
\item Description: The month of visiting the household
}
}
}
\usage{
R91
}
\description{
{R for Rural} \cr
{91 for 1391}
}
\keyword{datasets}
