% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91Data}
\alias{R91Data}
\title{Table R91Data}
\format{
A data frame with 22 columns:
\describe{
\item{Address}{"آدرس" Address}
\item{MahMorajeh}{"ماه مراجعه به خانوار" The month of visiting the household}
\item{ShoghlSarparast}{"شغل سرپرست خانوار" The job of the head of the household}
\item{NoeKhn}{"نوع خانوار (معمولي ساكن كد ،1 گروهي كد 2)" Household type (ordinary resident code 1, group code 2)}
\item{Takmil}{"بخش پ روي جلد پرسشنامه، سوال 19 (بله كد ،1 نه كد 2)" Questionnaire for substitute household(Yes cod 1, No code 2)}
\item{TakmilDescA}{"اطلاع جدول -2 علت حذف خانوار" Table-2 information on the reason for removing the household(
1.The family does not want to cooperate.
2.The family due to old age, mental disability, illiteracy of all family members or... is not able to answer.
\item{TakmilDescB}{"اطلاع جدول -3 علت جايگزيني خانوار" Table-3 information on the reason for replacing the household(
1.The household was absent during the census season.
2.The place is empty.
3.The place is not an ordinary residence for the household.
4.The place is destroyed, under construction or demolition.
5.The residence of the family is an institutional or non-resident.
6.The intend address was not found)}
\item{TakmilDescC}{"كد 7 جدول -3 ساير (با ذكر علت)" (Code 7, table-3, other (with the reason for replacing the household)}
\item{Jaygozin}{"اطلاع پرسش 20 (بله كد ،1 نه كد 2)"}
\item{JaygozinDescA}{"اطلاع جدول -4 علت حذف خانوار جايگزين به دلايل حذف" Table-4 the reason for the removal of the replacement household for reasons of removal)}
\item{JaygozinDescB}{"اطلاع جدول -5 علت حذف خانوار جايگزين به دلايل جايگزيني" Table-5 the re  ason for the removal of the replacement household for alternative reasons}
\item{JaygozinDescC}{"كد 7 جدول -5 ساير (با ذكر علت)" (Code 7, table-5, other (with the reason for the removal of the replacement household for alternative reasons)}
\item{BlkAbdJaygozin}{"شماره بلوك يا آبادي خانوار جايگزين (كه بايد با خانوار نمونهي اصلي يكسان باشد)" Block or settlement number of the alternative household (which must be the same as the original sample household)}
\item{RadifJaygozin}{"شماره رديف خانوار جايگزين در واحد نمونه گيري اوليه" Substitute household row number in the primary sampling unit}
\item{fStat}{Description of var15}
\item{fEnt}{Description of var16}
\item{fvfy}{Description of var17}
\item{user1}{Description of var18}
\item{user2}{Description of var19}
\item{user3}{Description of var20}
\item{user4}{Description of var21}
\item{Tekrari}{Description of var22}
}
}
\usage{
R91Data
}
\description{
R for Rural
91 for year of 1391
}
\keyword{datasets}
