% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91KarBargh1}
\alias{R91KarBargh1}
\title{Title for R91KarBargh1}
\format{
A data frame with 4 variables:
\describe{
\item{Address}{Address}
\item{DYCOL01}{Person row number}
\item{DYCOL02}{Activity row number}
\item{DYCOL03}{Income}
}
}
\usage{
R91KarBargh1
}
\description{
Title for R91KarBargh1
}
\keyword{datasets}
