% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{R91KarBargh2}
\alias{R91KarBargh2}
\title{Title for R91KarBargh2}
\format{
A data frame with 10 variables:
\describe{Income (employees in the agricultural sector)
\item{Address}{Address}
\item{DYCOL01}{Person row number}
\item{DYCOL02}{Job code}
\item{DYCOL03}{Type and feature of activities (agriculture)}
\item{DYCOL04}{Acreage}
\item{DYCOL05}{Gardening}
\item{DYCOL06}{Garden size}
\item{DYCOL07}{Animal husbandry}
\item{DYCOL08}{The number of small livestock}
\item{DYCOL09}{The number of big livestock}
\item{DYCOL10}{Poultry farming}
\item{DYCOL11}{Number of hens, chickens and roosters}
\item{DYCOL12}{Breeding fish}
\item{DYCOL13}{Pool size for breeding}
\item{DYCOL14}{Other agricultural activities}
\item{DYCOL15}{Description of var6 "تعمیرات خرمنکوب"}
\item{DYCOL16}{("تعمیرات تراکتور خریداری شده")}
\item{DYCOL17}{Description of var8}
\item{DYCOL18}{Description of var9}
\item{DYCOL19}{Description of var10}
\item{DYCOL20}{Description of var2}
\item{DYCOL21}{Description of var3}
\item{DYCOL22}{Description of var4}
\item{DYCOL23}{Description of var5}
\item{DYCOL24}{Description of var6}
\item{DYCOL25}{Description of var7}
\item{DYCOL26}{Description of var8}
\item{DYCOL27}{Description of var9}
\item{DYCOL28}{Description of var10}
\item{DYCOL29}{Description of var2}
\item{DYCOL30}{Description of var3}
\item{DYCOL31}{Description of var4}
\item{DYCOL32}{Description of var5}
\item{DYCOL33}{Description of var6}
\item{DYCOL07}{Description of var7}
\item{DYCOL08}{Description of var8}
\item{DYCOL09}{Description of var9}
\item{DYCOL10}{Description of var10}
}
}
\usage{
R91KarBargh2
}
\description{
Title for R91KarBargh2
}
\keyword{datasets}
