% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91}
\alias{U91}
\title{Table U91}
\format{
A data frame with 2 columns
\describe{
\item{Address:}{"آدرس"}{" Household address"}
\item{MahMorajeh:}{"ماه مراجعه به خانوار"} {"The month of visiting the household"}
}
}
\usage{
U91
}
\description{
U for Urban
91 for year of 1391
}
\keyword{datasets}
