% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91}
\alias{U91}
\title{Title for U91}
\format{
A data frame with 2 variables:
\describe{
\item{Address}{"آدرس" Address}
\item{MahMorajeh}{"ماه مراجعه به خانوار" The month of visiting the household}
}
}
\usage{
U91
}
\description{
Title for U91
}
\keyword{datasets}
