% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91P4S01}
\alias{U91P4S01}
\title{Title for U91P4S01}
\format{
A data frame with 16 variables:
\describe{"درآمد پولی اعضای شاغل خانوار از مشاغل مزد و حقوق بگیر" Get paid income from family members working in the wage
\item{Address}{"آدرس خانوار" Address}
\item{DYCOL01}{"شماره ردیف عضو شاغل" Job row number}
\item{DYCOL02}{"آیا در حال حاضر شاغل است؟" Is she/he currently employed?}
\item{DYCOL03}{"کد شغل" Job code}
\item{DYCOL04}{"فعالیت اصلی محل کار" Main activity of the workplace}
\item{DYCOL05}{"بخش عمومی 1 بخش تعاونی 2 بخش خصوصی 3" General Section 1 Cooperative Section 2 Private Section 3}
\item{DYCOL06}{"میزان ساعت های کار در روز" The amount of working hours per day}
\item{DYCOL07}{"تعداد روزهای کار در هفته" Number of working days per week}
\item{DYCOL08}{"مجموع درآمدهای ناخالص مستمر و غیرمستمر قبل از کسورات(ماه گذشته)" Total of continuous and non-continuous gross income before deficits(last month)}
\item{DYCOL09}{"مجموع درآمدهای ناخالص مستمر و غیرمستمر قبل از کسورات (12 ماه گذشته))" Total of continuous and non-continuous gross income before deficits(last 12 months)}
\item{DYCOL10}{"مزد و حقوق و مزایای مستمر (ماه گذشته)" Continuous wages and benefits (last Month)}
\item{DYCOL11}{"مزد و حقوق و مزایای مستر (12 ماه گذشته))" Continuous wages and benefits (last 12 Months)}
\item{DYCOL12}{"مزایای غیرمستمر (ماه گذشته)" Uncontinuous benefits (last month)}
\item{DYCOL13}{"مزایای غیرمستمر (12 ماه گذشته)" Uncontinuous benefits (last 12 months)}
\item{DYCOL14}{"مجموع درآمد خالص(ماه گذشته)" Total net income (last month)}
\item{DYCOL15}{"مجموع درآمد خالص (12 ماه گذشته)" Total net income (last 12 months)}
}
}
\usage{
U91P4S01
}
\description{
Title for U91P4S01
}
\keyword{datasets}
