% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documantation.R
\docType{data}
\name{U91P4S02}
\alias{U91P4S02}
\title{Title for U91P4S02}
\format{
A data frame with 3 columns:
\itemize{
\item \code{DYCOL01}:
Farsi name: \emph{آدرس}
Description: Household address
\item \code{DYCOL02}:
Farsi name: \emph{مالک خودرو}
Description: Car
}
}
\usage{
U91P4S02
}
\description{
Title for U91P4S02
}
\keyword{datasets}
